separate to grades.ts
