+++ server/src/routes/course.ts
@@ -2,30 +2,13 @@
 //
 // SPDX-License-Identifier: MIT
 
-import express, { Request, Router } from 'express';
-import multer, { FileFilterCallback, memoryStorage, Multer } from 'multer';
+import express, { Router } from 'express';
 
-import { addCourse, addGrades, getCourse } from '../controllers/course';
+import { addCourse, getCourse } from '../controllers/course';
 import { handleInvalidRequestJson } from '../middleware';
 import { controllerDispatcher } from '../middleware/errorHandler';
-import { ApiError } from '../types/error';
-import { HttpCode } from '../types/httpCode';
 
 export const router: Router = Router();
-
-const upload: Multer = multer({
-  storage: memoryStorage(),   // Store csv temporarily on the memory.
-  limits: {
-    fileSize: 1024 * 1024     // The max file size (in bytes), limit to 1 MB.
-  },
-  fileFilter: (_req: Request, file: Express.Multer.File, callback: FileFilterCallback) => {
-    if (file.mimetype == 'text/csv') {
-      callback(null, true);
-    } else {
-      return callback(new ApiError('incorrect file format, use csv format', HttpCode.BadRequest));
-    }
-  }
-});
 
 /**
  * @swagger
@@ -169,65 +152,4 @@ router.post(
   express.json(),
   handleInvalidRequestJson,
   controllerDispatcher(addCourse)
-);
-
-/**
- * @swagger
- * /v1/courses/{courseId}/instances/{instanceId}/grades/csv:
- *   post:
- *     tags: [Course]
- *     description: Add attainment points for users for specific course instance.
- *     requestBody:
- *       content:
- *         multipart/form-data:
- *     responses:
- *       200:
- *         description: Grading csv uploaded and parsed succesfully.
- *         content:
- *           application/json:
- *             schema:
- *               type: object
- *               properties:
- *                 success:
- *                   type: boolean
- *                   description: Success of the request.
- *       400:
- *         description: A validation or parsing error with the csv has occurred.
- *         content:
- *           application/json:
- *             schema:
- *               $ref: '#/definitions/Failure'
- *       401:
- *         description: The requester is not logged in.
- *         content:
- *           application/json:
- *             schema:
- *               $ref: '#/definitions/Failure'
- *       403:
- *         description: >
- *           The requester is not authorized to add grading data
- *           to the course instance (not teacher in the course).
- *         content:
- *           application/json:
- *             schema:
- *               $ref: '#/definitions/Failure'
- *       404:
- *         description: Course, course instance, or any of the attainable does not exist.
- *         content:
- *           application/json:
- *             schema:
- *               $ref: '#/definitions/Failure'
- *       409:
- *         description: >
- *           Course instance does not belong to the course or
- *           study attainment does not belong to the course instance.
- *         content:
- *           application/json:
- *             schema:
- *               $ref: '#/definitions/Failure'
- */
-router.post(
-  '/v1/courses/:courseId/instances/:instanceId/grades/csv',
-  upload.single('csv_data'),
-  controllerDispatcher(addGrades)
 );
